initial client TMDb, add 3rd party API, change movies (sequelize) popularity to INTEGER  Changes to be committed: - client related files 	new file:   TMDb/.expo-shared/assets.json 	new file:   TMDb/.gitignore 	new file:   TMDb/App.js 	new file:   TMDb/app.json 	new file:   TMDb/assets/icon.png 	new file:   TMDb/assets/splash.png 	new file:   TMDb/babel.config.js 	new file:   TMDb/package.json 	new file:   TMDb/yarn.lock - new migration, change model and seeder for movies (sequelize) 	new file:   server/movies/migrations/20200216000352-change-popularity-type-to-integer.js 	modified:   server/movies/models/movies.js 	modified:   server/movies/seeders/20200213101242-movies-seeder.js - 3rd API related files 	new file:   server/orchestrator/schemas/externalMovies.js 	modified:   server/orchestrator/schemas/movies.js 	modified:   server/orchestrator/schemas/tv.js

+++ server/movies/models/movies.js
@@ -15,7 +15,7 @@ module.exports = (sequelize, DataTypes) => {
       title: { type: DataTypes.STRING },
       overview: { type: DataTypes.STRING },
       posterPath: { type: DataTypes.STRING },
-      popularity: { type: DataTypes.DOUBLE },
+      popularity: { type: DataTypes.INTEGER },
       tags: { type: DataTypes.ARRAY(DataTypes.STRING) }
     },
     {}
